tween: Allow custom callbacks
In Camera, I'm making a custom container to handle the unconventional window
layout. This container could really make use of the Tween class, but it needs
to queue a resize rather than a redraw, as Tween does.

This commit allows a custom callback to be set, which will be called instead
of gtk_widget_queue_draw(). Tween will still ensure that the tick callback is
set and unset as necessary.

Also makes Tween public API. This will be useful in other places in Camera,
like rotating button icons.