Port Aperture to C
Porting the library to C allows finer control over the API and better
compatibility with other programming languages. Most GNOME libraries are
written in C, so Aperture should follow that standard.

A number of things have been removed in the port that were deemed unnecessary.
The code and API have been cleaned up in many places, although there is still
much work to do.

The documentation has also been improved and expanded.

The example project has been removed due to the API changes and the concerns
in #3. A new demo project will be added later.