viewfinder: Add camera-added handler
Previously, if there were no cameras plugged in and you plug one in, the
viewfinder would not notice. It would remain in the NO_CAMERAS state and you
wouldn't be able to use it again.

The new on_camera_added() handler checks whether the viewfinder is in the
NO_CAMERAS state, and if so, sets the viewfinder to use the new camera.