viewfinder: Async functions; redo error handling
- Make aperture_viewfinder_take_picture() and
aperture_viewfinder_stop_recording() async functions. This replaces the old
method, which was to emit signals when the operation was finished.

- aperture_viewfinder_take_picture() no longer saves the image to a file. It
only returns a pixbuf, and it is up to the application to save it.

- Redo error handling. The async functions now return errors when they fail,
instead of putting the viewfinder in the error state or doing nothing.

- Taking pictures is now done with a gdkpixbufsink rather than the image-capture
signal on the camerabin. This is simpler because it gives the viewfinder more
control over the process.

Fixes #7.