tests: Make ZBar-dependent tests skippable
This way, distributions that do not intend to make their Aperture package
depend on ZBar and its GStreamer plugin (although they should) do not have to
patch Aperture in order for the tests to pass.

In environments where ZBar should be a dependency and its absence would be an
issue, the `barcode_tests_skippable` option in Meson should be set to `false`.