Add ApertureCamera and ApertureDevice
ApertureCamera objects represent a camera connected to the system. They can be
obtained using an ApertureDeviceManager. This makes for a cleaner API, because
operations on a specific camera can be done with an ApertureCamera object
rather than on the device manager using an index.

It also paves the way for device-specific code using ApertureDevice. This is a
class that is meant to be subclassed for different devices, such as the
PinePhone or Librem 5, that Aperture might be used on. These implementations
will provide features such as camera flash and support for trickier camera
switching cases. The implementation is chosen at runtime based on things like
device tree name.

ApertureDevice is a private API. It is considered an implementation detail and
is not meant to be stable. ApertureCamera is mostly public, and applications
can use it to get information about the camera and do things like turn on/off
the flash. It also has some internal methods for working with GStreamer.